Intermediate release. Added support for ESP32 target. Updated ServoDispatchDirect to support servo easing. Updated AnalogMonitor to default to higher resolution on ESP32. Added TankDrive and Wifi classes for ESP32.

+++ src/ServoDispatch.h
@@ -137,7 +137,7 @@ public:
         _moveServosTo(servoGroupMask, startDelay, moveTime, moveTime, pos);
     }
 
-    void moveServosTo(uint32_t servoGroupMask, uint32_t startDelay, uint32_t moveTimeMin, uint32_t moveTimeMax, int16_t pos)
+    void moveServosTo(uint32_t servoGroupMask, uint32_t startDelay, uint32_t moveTimeMin, uint32_t moveTimeMax, float pos)
     {
         _moveServosTo(servoGroupMask, startDelay, moveTimeMin, moveTimeMax, pos);
     }
